feat(program): v0.5.9
Implement the dynamic-lender account layout refactor with u32 LTV fields across program, SDK parsers, and tests so contracts can scale to a borrower-selected cap up to 100 while preserving versioned compatibility checks.

+++ programs/stendar/src/instructions/pools.rs
@@ -1143,7 +1143,7 @@ pub fn pool_deploy_to_contract(ctx: Context<PoolDeployToContract>, amount: u64) 
     }
     if pool.min_ltv_bps > 0 {
         require!(
-            contract.ltv_floor_bps >= pool.min_ltv_bps,
+            contract.ltv_floor_bps >= pool.min_ltv_bps as u32,
             StendarError::PoolDeploymentRuleViolation
         );
     }
